models/functional_methods.py
from odoo import models, api
import os

class AudioManager(models.Model):
    _inherit = 'dudoxx.audio.recording'

    @api.model
    def transcribe_audio(self):
        # Logic to automate transcription via API
        pass

    @api.model
    def translate_transcription(self):
        # Logic to automate translation via API
        pass

    @api.model
    def chunk_audio(self):
        # Logic to split recordings into chunks
        pass

    @api.model
    def update_recording(self):
        # Logic to handle updates and versioning
        pass

    @api.model
    def upload_to_cloud(self):
        cloud_storage_url = os.getenv('CLOUD_STORAGE_URL')
        cloud_storage_key = os.getenv('CLOUD_STORAGE_KEY')
        cloud_storage_secret = os.getenv('CLOUD_STORAGE_SECRET')
        # Logic to manage cloud storage using the above credentials
        pass
