tests/test_functional_methods.py
from odoo.tests import TransactionCase

class TestFunctionalMethods(TransactionCase):

    def setUp(self):
        super(TestFunctionalMethods, self).setUp()
        self.AudioRecording = self.env['dudoxx.audio.recording'].create({
            'name': 'Test Recording',
            'duration': 120.0,
            'language_id': self.env.ref('base.lang_en').id,
            'transcription': 'This is a test transcription.',
        })

    def test_transcribe_audio(self):
        # Test the transcribe_audio method
        result = self.AudioRecording.transcribe_audio()
        self.assertIsNone(result)  # Replace with actual expected result

    def test_translate_transcription(self):
        # Test the translate_transcription method
        result = self.AudioRecording.translate_transcription()
        self.assertIsNone(result)  # Replace with actual expected result

    def test_upload_to_cloud(self):
        # Test the upload_to_cloud method
        result = self.AudioRecording.upload_to_cloud()
        self.assertIsNone(result)  # Replace with actual expected result
